+++ src/components/DatePicker/DatePicker.tsx
@@ -32,6 +32,8 @@ interface CustomTextInputProps extends InputProps {
   onClick?: (data: any) => any;
   /** The ReactDatePicker plugin has its own `required` prop so we use this to pass the value from the parent `DatePicker` component. */
   dsRequired?: boolean;
+  /** Whether or not to display the "Required"/"Optional" text in the label text. */
+  showOptReqLabel?: boolean;
 }
 
 // Main interface for the exported DS DatePicker component.
@@ -50,6 +52,8 @@ export interface DatePickerProps extends DatePickerWrapperProps {
   errored?: boolean;
   /** Adds the 'required' property to the input element(s). */
   required?: boolean;
+  /** Whether or not to display the "Required"/"Optional" text in the label text. */
+  showOptReqLabel?: boolean;
   /** Adds the 'disabled' property to the input element(s). */
   disabled?: boolean;
   /** Modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
@@ -80,6 +84,7 @@ const CustomTextInput = forwardRef<TextInputRefType, CustomTextInputProps>(
       labelText,
       disabled,
       dsRequired,
+      showOptReqLabel,
       errored,
       helperText,
       errorText,
@@ -95,6 +100,7 @@ const CustomTextInput = forwardRef<TextInputRefType, CustomTextInputProps>(
       labelText={labelText}
       disabled={disabled}
       required={dsRequired}
+      showOptReqLabel={showOptReqLabel}
       errored={errored}
       helperText={helperText}
       errorText={errorText}
@@ -147,6 +153,7 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
     errorText,
     errored,
     required,
+    showOptReqLabel = true,
     disabled,
     name,
     blockName,
@@ -163,6 +170,7 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
   // Both TextInput components share some props.
   let baseCustomTextInputAttrs = {
     dsRequired: required,
+    showOptReqLabel,
     // Always display the labels or the input fields when
     // the DatePicker component is a `dateRange` type.
     showLabel: dateRange ? true : showLabel,